split in more files. split in more functions. added work around for null point which can't be protobuffed
